+++ src/EntityGraphQL/Schema/BaseField.cs
@@ -156,50 +156,6 @@ public abstract class BaseField : IField
     }
 
     /// <summary>
-    /// To access this field all roles listed here are required
-    /// </summary>
-    /// <param name="roles"></param>
-    public IField RequiresAllRoles(params string[] roles)
-    {
-        RequiredAuthorization ??= new RequiredAuthorization();
-        RequiredAuthorization.RequiresAllRoles(roles);
-        return this;
-    }
-
-    /// <summary>
-    /// To access this field any role listed is required
-    /// </summary>
-    /// <param name="roles"></param>
-    public IField RequiresAnyRole(params string[] roles)
-    {
-        RequiredAuthorization ??= new RequiredAuthorization();
-        RequiredAuthorization.RequiresAnyRole(roles);
-        return this;
-    }
-
-    /// <summary>
-    /// To access this field all policies listed here are required
-    /// </summary>
-    /// <param name="policies"></param>
-    public IField RequiresAllPolicies(params string[] policies)
-    {
-        RequiredAuthorization ??= new RequiredAuthorization();
-        RequiredAuthorization.RequiresAllPolicies(policies);
-        return this;
-    }
-
-    /// <summary>
-    /// To access this field any policy listed is required
-    /// </summary>
-    /// <param name="policies"></param>
-    public IField RequiresAnyPolicy(params string[] policies)
-    {
-        RequiredAuthorization ??= new RequiredAuthorization();
-        RequiredAuthorization.RequiresAnyPolicy(policies);
-        return this;
-    }
-
-    /// <summary>
     /// Clears any authorization requirements for this field
     /// </summary>
     /// <returns></returns>