+++ src/EntityGraphQL/Schema/MethodField.cs
@@ -30,7 +30,7 @@ public abstract class MethodField : BaseField
         GqlTypeInfo returnType,
         MethodInfo method,
         string description,
-        RequiredAuthorization requiredAuth,
+        RequiredAuthorization? requiredAuth,
         bool isAsync,
         SchemaBuilderOptions options
     )