+++ src/EntityGraphQL/Schema/MutationType.cs
@@ -33,7 +33,7 @@ public class MutationType : ControllerType
         string? description,
         SchemaBuilderOptions? options,
         bool isAsync,
-        RequiredAuthorization requiredClaims,
+        RequiredAuthorization? requiredClaims,
         GqlTypeInfo returnType
     )
     {