+++ src/EntityGraphQL/Schema/SubscriptionField.cs
@@ -15,7 +15,7 @@ public class SubscriptionField : MethodField
         GqlTypeInfo returnType,
         MethodInfo method,
         string description,
-        RequiredAuthorization requiredAuth,
+        RequiredAuthorization? requiredAuth,
         bool isAsync,
         SchemaBuilderOptions options
     )